# public/developers/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Trustroots API",
        "description": "Trustroots.org API",
        "version": "0.2.0",
        "contact": {
            "name": "Trustroots",
            "url": "https://www.trustroots.org/developers/"
        }
    },
    "schemes": [
        "http",
        "https"
    ],
    "basePath": "/api",
    "produces": [
        "application/json",
        "text/html",
        "text/plain"
    ],
...
    "paths": {
...
        "/users": {
            "put": {
                "summary": "Update user profile.",
                "description": "Update user profile of currently logged in user.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User profile object.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Return updated user object.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/users/{username}": {
            "get": {
                "summary": "User by username.",
                "description": "Return user by username.",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "Username",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Return user object.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/users/avatar": {
            "post": {
                "summary": "Upload profile image.",
                "description": "Upload profile image for currently authenticated user.",
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "File object.",
                        "required": true,
                        "type": "file"
                    }
                ],
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Success. Returns image object."
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. Image is unable to be processed.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "salt": {
                    "type": "string"
                },
                "resetPasswordToken": {
                    "type": "string",
                    "description": "For reset password"
                },
                "resetPasswordExpires": {
                    "type": "string",
                    "format": "date-time",
                    "description": "For reset password"
                },
                "emailHash": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "format": "email"
                },
                "emailToken": {
                    "type": "string",
                    "description": "For email confirmations"
                },
                "emailTemporary": {
                    "type": "string",
                    "format": "email",
                    "description": "New email is stored here until it is confirmed"
                },
                "displayName": {
                    "type": "string",
                    "description": "This is (currently) generated in users.profile.server.controller.js"
                },
                "username": {
                    "type": "string",
                    "description": "Lowercase enforced username"
                },
                "displayUsername": {
                    "type": "string",
                    "description": "Unaltered original username"
                },
                "updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "birthdate": {
                    "type": "string",
                    "format": "date-time"
                },
                "locationFrom": {
                    "type": "string"
                },
                "locationLiving": {
                    "type": "string"
                },
                "newsletter": {
                    "type": "boolean",
                    "default": false
                },
                "public": {
                    "type": "boolean",
                    "description": "New users are public=false until they validate their email. If public=false, users can't email other users, can't be seen by other users. They are effectively black holed...\n"
                },
                "avatarUploaded": {
                    "type": "boolean"
                },
                "avatarSource": {
                    "type": "string"
                },
                "created": {
                    "type": "string",
                    "format": "date-time"
                },
                "roles": {
                    "type": "array",
                    "default": "user",
                    "items": [
                        {
                            "type": "string"
                        }
                    ],
                    "enum": [
                        "user",
                        "admin"
                    ]
                },
                "languages": {
                    "type": "array",
                    "items": [
                        {
                            "type": "string"
                        }
                    ]
                },
                "gender": {
                    "type": "string",
                    "default": "",
                    "enum": [
                        "",
                        "male",
                        "female",
                        "other"
                    ]
                },
                "description": {
                    "type": "string"
                },
                "tagline": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "additionalProvidersData": {
                    "type": "object"
                },
                "providerData": {
                    "type": "object"
                },
                "provider": {
                    "type": "string",
                    "description": "Internal. Relates to oauth logins. Currently only accepted auth provider is 'local'."
                },
                "seen": {
                    "type": "string",
                    "format": "date-time",
                    "description": "The last time the user was logged in (uncertain if its live right now 5 Apr 2015)"
                }
            },
            "required": [
                "email",
                "username",
                "provider"
            ]
        },
...
        "Error": {
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}